Drop breadcrumb script
I'll be honest, looking back this script was completely unnecessary. It was used to make the breadcrumb shrink in size, destroy the breadcrumb after a certain amount of time and tell the player they can place another. But all of these could have been done from the player controller script, all noted for future reference.