End reach script
This just made it so when the player collides with the object at the end of the level, the player moves to the next. Again this could have been merged into the character controller script. Instead of using the collider script of the end object, it could instead use the collider script of the player and check if the object collided with is the end object.