Level change script
This script was what controlled the new levels being created. A very complicated way that makes me confused how we even got to this outcome. But I don't feel we needed all this for the script, I don't exactly even know what is happening in the update function, but we could have definitely done without it. Leaving just the script which chooses what level we are on. This could have been done in the game master script which we did have, and have it be a function instead that runs when the player collides with the end game object.